Initial commit of configuration files for developing via vscode.
Neither of these files should be released, they are for configuring development environment only. The file .npmignore specifies what NOT to release.

.npmignore
This tells npm what not to release - it is identical to .gitignore, with the addition of the .vscode folder. NPM looks for .npmignore first, and if this does not exist, it then looks for .gitignore. So we want this file to contain everything in .gitignore, plus the .vscode directory.

(The alternative is to whitelist what should be released in the package via "files" {} from within package.json.

settings.json:
Configures the mocha sidebar for running index.test.js via vscode mocha testing sidebar extension, which will not work without the mocha.sideBarOptions and mocha.Options blocks. unfortunately the sidebar doesn't pick up the package.json mocha script settings.

launch.json:
has two profiles for debugging via vscode. The first is for launching node with a script named "test.js" (which doesn't exist, btw) and the second starts mocha with the index.test.js from within the debugger. The flag stopOnEntry means that execution will pause at the entry point, as if there was a breakpoint set on the first line of code.

Mocha can be launched in the debugger using the "Run script: test" which uses the pre-defined settings from package.json scripts block. However I find having this additional profile available helpful for fine-tuning. And Im committing it incase I forget how to do it.